refactor: KboGameResult -> KboGameSummaryResult 로 수정
