+++ backend/src/main/java/com/yagubogu/game/domain/ScoreBoard.java
@@ -1,0 +1,4 @@
+package com.yagubogu.game.domain;
+
+public class ScoreBoard {
+}
